+++ crates/ryo3-jiff/src/ry_zoned.rs
@@ -288,7 +288,7 @@ impl RyZoned {
 
     #[pyo3(
         warn(
-            message = "`intz` is deprecated, use `in_tz` instead",
+            message = "`ZonedDateTime.intz` is deprecated; use `ZonedDateTime.in_tz` instead [removal: v0.0.93]",
             category = pyo3::exceptions::PyDeprecationWarning
         )
     )]